% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checker.R, R/print.R
\name{edns_test}
\alias{edns_test}
\alias{print.edns_check_output}
\title{RFC 6891 (EDNS) Compliance Checker}
\usage{
edns_test(zone)

\method{print}{edns_check_output}(x, ...)
}
\description{
Uses \href{https://dnsflagday.net/}{the back end of the interactive compliance checker}
to test DNS servers for a given domain. Printing the resultant object will show
the status of the input domain and auto-browse (in interactive sessions) to the
generated compliance report. Pass \code{browse = FALSE} to \code{print()} to suppress
opening the report URL.
}
\examples{
edns_test("rstudio.com")
}
\references{
\url{https://ednscomp.isc.org/}; \url{https://tools.ietf.org/html/rfc6891}
}
\keyword{internal}
