% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ednstest-package.R
\docType{package}
\name{ednstest}
\alias{ednstest}
\alias{ednstest-package}
\title{'RFC' 6891 ('EDNS') Compliance Checker}
\description{
Checks 'Domain Name System' ('DNS') zones for compliance with
'RFC' 6891 ('EDNS').
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ednstest}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ednstest/issues}
}
}
\references{
\url{https://ednscomp.isc.org/}; \url{https://tools.ietf.org/html/rfc6891}
}
\author{
Bob Rudis (bob@rud.is)
}
