% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ednstest-package.R
\docType{package}
\name{ednstest}
\alias{ednstest}
\alias{ednstest-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ednstest}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ednstest/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
